AngelScript updated to rev. 2252.
git-svn-id: https://cvet.by:8443/svn/FOnlineSource/Main@1007 c2c6a35e-ad2b-4e1f-bb64-97b989be360b

Former-commit-id: ce315f50bb3a71f56dbf6eb04e4f6f80119538ba

+++ Source/AngelScript/sdk/add_on/scriptmath/scriptmath.cpp
@@ -111,7 +111,7 @@ bool closeTo(float a, float b, float epsilon)
 	if( a == b ) return true;
 
 	// When very close to 0, we can use the absolute comparison
-	float diff = fabs(a - b);
+	float diff = fabsf(a - b);
 	if( (a == 0 || b == 0) && (diff < epsilon) )
 		return true;
